refactor contracts manager; deployment; blockchain client
