+++ dapps/tests/app/old_embark.json
@@ -1,5 +1,5 @@
 {
-  "contracts": ["app/contracts/**"],
+  "contracts": ["app/contracts/**", "contracts/**"],
   "app": {
     "js/jquery.js": "app/js/_vendor/jquery.min.js",
     "js/bootstrap.js": "app/js/_vendor/bootstrap.min.js",
